Set Gate PPSPM, Bendahara, Admin

+++ app/Http/Controllers/MapingppkController.php
@@ -6,6 +6,7 @@ use App\Models\pagu;
 use App\Models\User;
 use App\Models\mapingpaguppk;
 use App\Models\mapingstafppk;
+use Illuminate\Support\Facades\Gate;
 
 class MapingppkController extends Controller
 {
@@ -16,6 +17,10 @@ class MapingppkController extends Controller
      */
     public function index()
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
         return view('referensi.maping_ppk.index',[
             'data'=>User::pegawaisatker()->ppk()->get()
         ]);
@@ -29,6 +34,14 @@ class MapingppkController extends Controller
      */
     public function showpagu(User $ppk)
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
+        if($ppk->satker != auth()->user()->satker){
+            abort(403);
+        }
+
         return view('referensi.maping_ppk.pagu.detail',[
             'data'=>$ppk->paguppk,
             'ppk'=>$ppk
@@ -37,6 +50,14 @@ class MapingppkController extends Controller
 
     public function showstaf(User $ppk)
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
+        if($ppk->satker != auth()->user()->satker){
+            abort(403);
+        }
+
         return view('referensi.maping_ppk.staf.detail',[
             'data'=>$ppk->stafppk,
             'ppk'=>$ppk
@@ -51,6 +72,14 @@ class MapingppkController extends Controller
      */
     public function editpagu(User $ppk)
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
+        if($ppk->satker != auth()->user()->satker){
+            abort(403);
+        }
+
         return view('referensi.maping_ppk.pagu.update',[
             'data'=>pagu::pagunonppk()->get(),
             'ppk'=>$ppk
@@ -59,6 +88,14 @@ class MapingppkController extends Controller
 
     public function editstaf(User $ppk)
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
+        if($ppk->satker != auth()->user()->satker){
+            abort(403);
+        }
+
         return view('referensi.maping_ppk.staf.update',[
             'data'=>User::stafnoppk()->get(),
             'ppk'=>$ppk
@@ -74,6 +111,18 @@ class MapingppkController extends Controller
      */
     public function updatepagu(User $ppk, pagu $pagu)
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
+        if($ppk->satker != auth()->user()->satker){
+            abort(403);
+        }
+
+        if($ppk->satker != $pagu->kodesatker){
+            abort(403);
+        }
+
         mapingpaguppk::create([
             'pagu_id'=>$pagu->id,
             'user_id'=>$ppk->id
@@ -83,6 +132,18 @@ class MapingppkController extends Controller
 
     public function updatestaf(User $ppk, User $staf)
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
+        if($ppk->satker != auth()->user()->satker){
+            abort(403);
+        }
+
+        if($ppk->satker != $staf->satker){
+            abort(403);
+        }
+
         mapingstafppk::create([
             'staf_id'=>$staf->id,
             'ppk_id'=>$ppk->id
@@ -98,12 +159,28 @@ class MapingppkController extends Controller
      */
     public function destroypagu(User $ppk, mapingpaguppk $mapingppk)
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
+        if($ppk->satker != auth()->user()->satker){
+            abort(403);
+        }
+
         $mapingppk->delete();
         return redirect('/maping-ppk/'.$ppk->id.'/pagu')->with('berhasil', 'Pagu Berhasil Di Hapus dari PPK');
     }
 
     public function destroystaf(User $ppk, mapingstafppk $mapingstafppk)
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
+        if($ppk->satker != auth()->user()->satker){
+            abort(403);
+        }
+
         $mapingstafppk->delete();
         return redirect('/maping-ppk/'.$ppk->id.'/staf')->with('berhasil', 'Pagu Berhasil Di Hapus dari PPK');
     }
