+++ app/Http/Controllers/MapingstafppkController.php
@@ -6,6 +6,7 @@ use App\Models\unit;
 use App\Models\User;
 use Illuminate\Http\Request;
 use App\Models\mapingunitstafppk;
+use Illuminate\Support\Facades\Gate;
 
 class MapingstafppkController extends Controller
 {
@@ -16,6 +17,10 @@ class MapingstafppkController extends Controller
      */
     public function index()
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
         return view('referensi.maping_staf_ppk.index',[
             'data'=>User::pegawaisatker()->stafppk()->get()
         ]);
@@ -50,6 +55,14 @@ class MapingstafppkController extends Controller
      */
     public function showunit(User $stafppk)
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
+        if($stafppk->satker != auth()->user()->satker){
+            abort(403);
+        }
+        
         return view('referensi.maping_staf_ppk.unit.detail',[
             'data'=>$stafppk->unitstafppk,
             'stafppk'=>$stafppk
@@ -64,6 +77,14 @@ class MapingstafppkController extends Controller
      */
     public function editunit(User $stafppk)
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
+        if($stafppk->satker != auth()->user()->satker){
+            abort(403);
+        }
+
         return view('referensi.maping_staf_ppk.unit.update',[
             'data'=>unit::Nostafppk($stafppk->nip)->get(),
             'stafppk'=>$stafppk
@@ -79,6 +100,14 @@ class MapingstafppkController extends Controller
      */
     public function updateunit(User $stafppk, unit $unit)
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
+        if($stafppk->satker != $unit->kodesatker){
+            abort(403);
+        }
+
         mapingunitstafppk::create([
             'user_id'=>$stafppk->id,
             'unit_id'=>$unit->id
@@ -94,6 +123,10 @@ class MapingstafppkController extends Controller
      */
     public function destroyunit(User $stafppk, unit $unit)
     {
+        if (! Gate::allows('admin_satker', auth()->user()->id)) {
+            abort(403);
+        }
+
         $stafppk->unitstafppk()->detach($unit->id);
         return redirect('/maping-staf-ppk/'.$stafppk->id.'/unit')->with('berhasil', 'Unit Berhasil Di Hapus dari Staf PPK');
     }
